+++ utils.py
@@ -101,7 +101,7 @@ def preprocess_image(im):
 
     I = np.flip(I, 2)  # BGR
     I = I - [[[104.00698793, 116.66876762, 122.67891434]]]  # subtract mean - whitening
-    I = np.reshape(I, (1,) + I.shape)
+    I = np.reshape(I, (1,) + I.shape).astype(np.float32)
     return I
